services/pulse_discord_bot/requirements.txt
discord.py==2.3.2
httpx==0.27.0
redis==5.0.8
aiohttp==3.9.5
